Fix | Escape return when payment is paid

+++ app/Http/Controllers/Pembayaran/RegistrasiController.php
@@ -91,6 +91,8 @@ class RegistrasiController extends Controller
         $user = auth()->user();
         $data = Pembayaran::where('user_id', $user->id)->where('kategori', 'registrasi')->where('status', '!=', 1)->first();
 
+        if(is_null($data)) return response()->redirectToRoute('biodata.create');
+
         if(date('Y-m-d H:i:s',strtotime($data->expiredDate)) <= date('Y-m-d H:i:s')){
             $data->delete();
             return $this->makeBNIInvoice();
